# api.yml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: Microservicio para listas
  description: Python + PostgreSQL
  license:
    name: Software Architecture 2019-I
host: 192.168.99.101:4004
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /lists-ms/lists/{id}:
    get:
      summary: Retorna las Lista de listas segun el id de usuario
      operationId: lists_by_user
      parameters:
        - name: id
          in: path
          description: El id del usuario
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/lists'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /lists-ms/list_all/{id}:
    get:
      summary: Retorna una lista, incluyendo los datos del object list, con todos los lugares a visitar
      operationId: list_with_places_by_list
      parameters:
        - name: id
          in: path
          description: El id de la lista
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/list_places'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /lists-ms/places/{id}:
    get:
      summary: Retorna una lista, con todos los lugares a visitar
      operationId: places_by_list
      parameters:
        - name: id
          in: path
          description: El id de la alista
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/places'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
definitions:
  list:
    type: object
    properties:
      id:
        type: integer
      id_user:
        type: integer
      name:
        type: string
      comment:
        type: string
      estimatedDate:
        type: string
        format: date
      order:
        type: integer
  place:
    type: object
    properties:
      id:
        type: integer
      id_list:
        type: integer
      placeLon:
        type: number
      placeLat:
        type: number
  lists:
    type: object
    required:
      - id_ser
    properties:
      lists:
        type: array
        items:
          $ref: '#/definitions/list'
  list_places:
    type: object
    properties:
      list:
        $ref: '#/definitions/list'
      places:
        type: array
        items:
          $ref: '#/definitions/place'
  places:
      type: object
      properties:
        places:
          type: array
          items:
            $ref: '#/definitions/place'

# Errors
  ErrorModel:
    type: object
    required:
      - message
      - code
    properties:
      message:
        type: string
      code:
        type: number
      description:
        type: object